+++ src/Parse.hs
@@ -40,7 +40,7 @@ import Parse.Type (TibetScript (..), TibetSyllable (..), TibetWylieMap, WylieScr
 import Parse.WylieWord (wylieWord)
 import Type (HibetError (..))
 
-import Control.Monad.Except (Except, liftEither)
+-- import Control.Monad.Except (Except, liftEither)
 import qualified Data.HashMap.Strict as HM
 import Data.RadixTree (RadixTree, fromFoldable_, lookup)
 import Data.Text (Text)
@@ -56,13 +56,12 @@ import Prelude hiding (lookup)
 -- TODO: Check perfomance on big input
 toTibetan
     :: WylieTibetMap
-    -> Except HibetError [WylieScript]
-    -> Except HibetError [TibetScript]
-toTibetan bi ts = do
-    txt <- ts
+    -> [WylieScript]
+    -> Either HibetError [TibetScript]
+toTibetan wtMap wList = do
     let look :: WylieSyllable -> Maybe TibetSyllable
-        look = flip HM.lookup bi
-    liftEither $ flip traverse txt $ \case
+        look = flip HM.lookup wtMap
+    flip traverse wList $ \case
       ScriptWylie w -> case look w of
         Nothing  -> Left NotFound
         Just res -> Right $ ScriptTibet res
@@ -71,13 +70,12 @@ toTibetan bi ts = do
 -- | Convert parsed tibetan text to wylie.
 toWylie
     :: TibetWylieMap
-    -> Except HibetError [TibetScript]
-    -> Except HibetError [WylieScript]
-toWylie hm ets = do
-    ts <- ets
+    -> [TibetScript]
+    -> Either HibetError [WylieScript]
+toWylie twMap tList = do
     let look :: TibetSyllable -> Maybe WylieSyllable
-        look = flip HM.lookup hm
-    liftEither $ flip traverse ts $ \case
+        look = flip HM.lookup twMap
+    flip traverse tList $ \case
       ScriptTibet w -> case look w of
         Nothing  -> Left NotFound
         Just res -> Right $ ScriptWylie res
@@ -125,15 +123,15 @@ Something won't be consumed, and dropped.
 At second, current parsers will be refactored to more and more correct result against flat list.
 
 -}
-parseWylieInput :: RadixTree () -> Text -> Except HibetError [WylieScript]
+parseWylieInput :: RadixTree () -> Text -> Either HibetError [WylieScript]
 parseWylieInput radix txt  = do
-    ls <- parseExcept wylieWord txt
+    ls <- parseEither wylieWord txt
     pure $ map (lookupWylieScript radix) ls
 
 -- | Parse text to tibetan or fail.
-parseTibetanInput :: RadixTree () -> Text -> Except HibetError [TibetScript]
+parseTibetanInput :: RadixTree () -> Text -> Either HibetError [TibetScript]
 parseTibetanInput radix txt  = do
-    ts <- parseExcept tibetanWord txt
+    ts <- parseEither tibetanWord txt
     pure $ map (lookupTibetScript radix) ts
 
 -- | Make wylie radix tree from syllables.